Add animations to scorpion boss, finalise level one and two terrain, add poison effect and change to new trees

+++ Assets/Scripts/Enemy/HealthBar.cs
@@ -13,7 +13,7 @@ namespace Enemy
       slider.value = maxHealth;
     }
 
-    public void SetCurrentHealth(int currentHealth)
+    public void SetCurrentHealth(float currentHealth)
     {
       slider.value = currentHealth;
     }
